[AllStrainContainer]  [4:06pm]
Type '{}' is not assignable to type 'string'.ts(2322)

export default  function AllStrainContainer(page:string){    

wanted to see if i could treat this component as an exported pure function with string argument parameters 
if so, create a string argument that parses for commands based on $('document') or [window.location] URL of the [ https://www. ... .com ]

This creates an error in the index.tsx page.