[AllStrain] toggle bg btn works [12:04pm]
state created and toggle button working.
state based rendering.

i wasn't sure if the map would work but it does. 
the background-repeat is still applying to both bg. The old style doesn't have the bg like that. will have to toggle classes as well.

now realizing probably don't have to get rid of the unordered list.
can just toggle the class based on state and for the new design have there be no class for the unordered list like it doesn't exist.