[AllStrain] instead of 0.0 state {nothing} [12:25pm]
kind of funny. if you do:
<li> '' </li>
you will see '' rendered onto the page.

if you do 
const [nothing, setNothing] = useState()
👍 <li> {nothing} </li> 

there is nothing on the page.

can also make opacity: 0.0 or 1.0 on hover. 

few different things to do. Starting to like idea of leaving names off of this container until you save it and it tells user/human which data.strain/name  saved.
more in line with the theming of digging and finding gold in a mine