siblings() because .attr going on sister elem [9:34pm]
while trying to use the modularized jquery based function i made that takes a [typeof object] and performs the .attr() method parsing for user [property, value]
i realized that i'm not trying to change the input, the expression that is being applied setting a value attribute onto a div element which will be doing nothing. 

creating a modularized function that accepts 1 (typeof object) input that returns the jq siblings() method and the data returned from invoking it.