invoke with event.target instead of event.target.innerText [10:25pm]
allow the object to be for when the request is the object of the strain being sent import/invocation style

that typeof separating functionality works. this above line of console.log() is returned because i switched to getSpecifiedStrain(target) which is the 
event object.target

also invoking the function is working with console.log(). hitting it with the post route hides the console.logs()