[Container] Axios.Create transformResponse [10:00am]
if prisma could be run in the browser it would be very easy to use:
pure-function with passed params invocation, clean-up request data with functionality, parse & pass response.

prisma cant be run in browser and sending the data in XML or Http based route-hitting methods is creating a URL like so:
localhost:3000/api/getSpecifiedStrain?json=dataFromAjaxOrAxios

looking to use lastIndexOf( ' ? ' ) to grab the string that starts after the last question marks and use regex to slice json= off of that 

thatll give the {strain.strain} data to send response data back with. 

almost purposely making this more difficult because the APIcall() is up and running and could be a very easy way to bypass all of this and make only the necessary db.prisma.table.create() / db.prisma.table.findMany/Unique() functionality.