working with data np but with oversight [11:25pm]
kind of tangled up:
human error: 
as I was entering code that got returned. only the strainId didn't get returned. there is no strainId in the API, only in the database.

i realized that the loop/forEach { i + 1 } would correspond to its ID but then we have a duplicate of white widow from the[ 'api/allStrain post ] 
that i was going to figure out later. the duplicate will throw off ids being accurate after. 
Could also set up a function that loops against each other and uses the master API list as a way of validating and updating the IDs within a conditional that checks for it


might figure that out now and use this index trick as solution