[StrainDisplay] alternative display based on accident [3:57pm]
didn't think of this first, but react in-line styling on the bootstrap element is overriding the built-in-bootstrap-defaults 
issue posted: i kind of like the look of the accident.

one could make a toggleable state and have either display be built-in to the changebg toggle that displays the new or old design of the app.

props based CSS in styled components will allow this. just have to create the state.