[AllStrainContainer] JSON.parse() [10:03am]
i was going to post an issue but was embarassed to do so.

My problem was that my JSON was coming back as a string. realized there was something to learn because that sounded like an obvious fact.
[ JSON.parse() or .JSON() ]  basically toggles  data="json-string" from the data.endpoints that we now have the returned data call declared as

i saw the string data in the console and it seemed to not look like the object I'd need.
went on to kind of kept working past that without triple checking.