whats going on more easily understood now [9:48pm]
changed param 'element' to api (i was originally going to create state and send over object-endpoint state from $('') jquery targeting the display div container.

theres an additional parameter called length. hoping to not need this but:
the length will help to restart when we reach the end of the api. array-indexing may make need for [index + 1] (and will probably need state to do so) 

thats why i was going to pass an element through as well, to get the innerText of that element and continually loop through the API and do [i + 1] 
whenever the loop reaches the current active innerText so it progresses onto the next endpoint.