structure of expression, invocation, params [10:48am]
params 
errorobj is from catch { params } object,
errorState is the state (might not need)
setErrorState(to set the state from within the export function)

wondering whether runtime for the expression is longer by returning '404' and invoke function like so:

let returnedError = await ERROR( {...} ) // 404

or if changing the state externally is faster. I would believe that returning just data and changing state in local component is faster but good to know both

*****
might not need the url because the url has to do with the try block and this expression is invoked only within the catch block.