h2 elem instead of p [3:35pm]
thought of using stripe API as if you could pay for the stuff on this site and send a message that you can't.
was glad to integrate Google API and will probably do google API again. using stripe without actually taking payments. 
it'd be nice to return data from their objects