almost posted issue. redux config [8:08pm]
Loading...
Expected 0 arguments, but got 1.ts(2554)

 # const dispatch = useDispatch() 
// almost posted issue. the storing of the invocation of useDispatch method was done in a line of code by the imports in the top level block and broke app

    let parents:string = props.redux   
    let parent1 = props.parent1 
    let parent2:string = props.parent1 

was also having trouble because i was doing this:

        // reduxparents = dispatch( { type: "SET_PARENTS_1", payload: { parents: newstr[0]}}

i was using props as state which now as i type shouldn't seem to be problematic. you're not supposed to make dispatches of state that is passed as props

global redux state is to avoid passing props. but have also successfully passed the redux state as props from parent to this container component.