ReturnWrong [5:54pm]
still some duplicates sometimes sneaking-in.

i added another layer of double checking based on the other function parameter that is the :string[] arr of [strain.name] added to the list when they become 
intended wrong guesses

                } else {                    
                    if (dontuse.includes(strain.strain)) {                                     // this code right here
                        let newstrain = await Random(strains)
                        let name:string = newstrain.strain
                        stringbucket.push(name)
                    } else { stringbucket.push(strain.strain) }                                                 
                }

// instead of using the .includes() in that code right there.

1) i first started with forEach()

but then the:
2) dontuse:string|string[]

even though there is a string array type definition on that dontuse function parameter 
can't use forEach() on an array typed like that. or atleast i didn't get it to work.

instead of forEach()... .includes() check for existence

this could still be beat though but there's much less chance at that.

With no locks on this for some reason... almost 3/4 times before the [guess #1 and guess #2] always become the same strain

destroys the decency of the game.

*** quick 2 minute update: tried 12x to get a duplicate value out of the values appended to the labels and couldn't.
its still not foolproof though. 

the game restarts and the non-duplicating + generation of entirely new values stands.