NumberSearch [11:09am]
iterates over the scrambled array vs the not-scrambled array.
now when you press a number after clicking onto the magnifying glass 

(which activates the receptivity of the search input):

when you type a number into the input it returns a random 3 set of strains. 
i did this so things felt a little less robotic.