2 hour issue #65
i was checking and toggling state with strict equality comparison to the negative:
let input = onChange.value

1)
input.replace()
perform regex that removed certain characters

2)
check length. Once the regex is performed, if the constraint/condition has been met, The string returned will be void of the character that passes the condition
once that string is less than the original string that doesn't have a regex performed upon it, it activates the condition toggling the state
if (inputAfterRegex.length < input.length )  

problem is that:
special characters and numbers were interfering with surviving the escape regex, causing an offset in the length without the condition truly being met.
Which would improperly tag the InputAfterRegex.length to be less than when it deserved to be the same as the original input.length which would not survive the if statement and satisfy the else block keeping state untoggled and the text unhighlighted.

lesson in refactoring:
using .match() and looking only for the character that passes the condition
and checking for the existence of that character giving any length to (inputAfterMatch)

all you have to do is check for the existence of length as if it were a boolean toggling from false to true:

  let specialRegex = actualstring.match(specialPattern)

                if (specialRegex) {
                    specialcharset('true')
                } else {
                    specialcharset('false')
                }
               
also part of the problem untangling that issue was that:

the number regex will take the characters that are numbers out of the original string. it worked perfectly no problem.
but if you type a number as the first character of the input,

the number survives the regex and the length is 1,
where as if you typed (b1) into the input...
the regex would take the b off returning a length of 1, and the original input (b1) would have a length of 2 satisfying the condition performed by assessing 
the negative/opposite to checking the existence

match solved everything
[12:54pm]