remove checkbox checking, persist without bias [8:45pm]
before these expressions were within conditional logic connected to the checked checkbox.

The problem with this is that:          (say for instance):
We've typed 'backwardsHat!459@gmail.com' into the email input.

this password will pass the requiring conditions of: [having a] -> 1) UPPERcase character, 2) !special character, 3) number...

but if the checkbox that was checked while creating that very password with the password input was the 'email checkbox'
none of those conditions will show up as validating the password requirements which makes things feel 

fake / plastic.

removing the barriers and the assertations persist independently of whichever checkbox is checked.

doing so is in place of:

changing the also kind of nuisance of being able to type on password input while checkbox is checked. If checkbox followed inputs as well that'd be nice.

the checkboxes look pretty nice though so forcing user to click on them brings out significance imo.

Letting the checkboxes and state and highlighted text indicating passing validations makes a huge difference. 

[8:50pm]